requirements.txt
python-telegram-bot==20.7
PyGithub==2.1.1
paramiko==3.4.0
